server/requirements.txt
aiohttp
aiohttp_cors
gunicorn
lxml
pandas
xlrd
xlwt
pytest-aiohttp
python-slugify
Sastrawi
